pyproject.toml
[project]
name = "parser"
version = "0.1.0"
description = ""
authors = [
    {name = "Alejandro Alvarez Espejo"}
]
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "pyside6 (>=6.9.1,<7.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pyinstaller (>=6.14.2,<7.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "odfpy (>=1.4.1,<2.0.0)",
    "striprtf (>=0.0.29,<0.0.30)",
    "ebooklib (>=0.19,<0.20)",
    "markdown (>=3.8.2,<4.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "deep-translator (>=1.11.4,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
